UBCompassTool upgrade
Added new functionality to the compass.

1. Position and size normalization parameters.
These parameters allow to align the position and size of the compass to the gridSize. Parameters are only enabled when the pageBackground is crossed.

2. Right angle rotation.
This functionality allowsu to turn the compass counterclockwise to the nearest degree multiple 90 (0, 90, 180, 270).

3. Rotation of the compass by a given angle.
You can rotate the compass by a given number of degrees absolutely or relative to the current rotation of the compass. At the same time, you can draw (this is optional).